+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog for Bitmark SDK for Swift
 
+## 2.6.0
+### Breaking changes:
+Bitmark Swift SDK v2.6.0 introduces some breaking changes to align with other Bitmark SDKs:
+- Remove `Account.parseAccountNumber(accountNumber:)` in favor of `Account.validateAccountNumber(accountNumber:)`.
+- Remove `AccountNumber.parse()` in favor of `AccountNumber.validate()`
+- `Transfer.respond(withResponseParams:)` will return an optional String as txId when reponding with acceptance.
+
 ## 2.0.0
 ### Features:
 - Account: create new account, generate Recovery Phrase, recover account from Recovery Phrase